pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*", "models*", "rules*", "scripts*"]

[project]
name = "neuralwarden"
version = "2.0.0"
description = "AI NeuralWarden Pipeline with Multi-Model Routing"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "pydantic>=2.0",
    "pinecone-client>=5.0.0",
    "langchain-pinecone>=0.2.0",
    "gradio>=5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
neuralwarden = "main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
